Add Skin Tan Never Changes mod for USA v1.0 and PAL
This mod makes it so your skin tan never gets lighter or darker -
intended to be used in tandom with save editors to permanently change
your skin color.

In the future, I'd like to implement this in a more robust way, with a
whole bit of custom dialog with Kapp'n at the beginning to choose your
skin color, but that's gonna be some work lol.